test: actually type-check the generics assertions
The expectTypeOf assertions in generics.spec.ts were never verified. They
are a no-op at runtime, and nothing in the pipeline ran tsc over test/, so
a deliberately wrong assertion like expectTypeOf(false).toBeNumber() passed
with 44/44 green.

Enabling Vitest's typecheck is not sufficient on its own either: its default
include only matches *.{test,spec}-d.ts, so *.spec.ts files are silently
skipped, and it falls back to the root tsconfig, whose "node10" resolution
cannot read the ESM-only type exports of Vitest 4 / Vite 8.

- rename generics.spec.ts to generics.spec-d.ts so it is picked up, and drop
  the try/catch/finally + expect(true).toBeTruthy() scaffolding, which
  asserted nothing at runtime
- enable typecheck in vitest.config.e2e.ts, pointed at test/tsconfig.json
- set moduleResolution "bundler" for the test tsconfig

Verified by mutation: a wrong expectTypeOf, a removed @ts-expect-error, and
a return-type regression on EventBus#publish each fail the e2e run.
